ValidationCompeleted

+++ HR_Management.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR_Management.Application.Exceptions;
 using HR_Management.Application.Features.LeaveAllocations.Requests.Commands;
 using HR_Management.Application.Persistance.Contracts;
+using HR_Management.Domain;
 using MediatR;
 
 namespace HR_Management.Application.Features.LeaveAllocations.Handlers.Commands;
@@ -23,6 +25,9 @@ public class DeleteLeaveAllocationCommandHandler:IRequestHandler<DeleteLeaveAllo
     public async Task Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
     {
         var leaveAllocation =  await _leaveAllocationRepository.GetByIdAsync(request.Id);
+        if (leaveAllocation == null)
+            throw new NotFoundException(nameof(LeaveAllocation), request.Id);
+
         await _leaveAllocationRepository.Delete(leaveAllocation);
     }
 }
