+++ HR_Management.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR_Management.Application.Exceptions;
 using HR_Management.Application.Features.LeaveTypes.Requests.Commands;
 using HR_Management.Application.Persistance.Contracts;
+using HR_Management.Domain;
 using MediatR;
 
 namespace HR_Management.Application.Features.LeaveTypes.Handlers.Commands;
@@ -23,6 +25,9 @@ public class DeleteLeaveTypeCommandHandler:IRequestHandler<DeleteLeaveTypeComman
     public async Task Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveType = await _leaveTypeRepository.GetByIdAsync(request.Id);
+        if (leaveType == null)
+            throw new NotFoundException(nameof(LeaveType),request.Id);
+        
         await _leaveTypeRepository.Delete(leaveType);
 
     }
